actions

+++ neural/generate_measures_action_dataset.py
@@ -57,7 +57,7 @@ with torch.no_grad():
         with ctx:
             actions = model.encode_actions(batch, bpm, force_manual=True, force_transfer=False)
         
-        print(((arr[i*batch_size:(i+1)*batch_size] - actions.float().cpu().detach().numpy().astype(np.float16)) ** 2).mean().item())
+        print(((arr[i*batch_size:(i+1)*batch_size].astype(np.float32) - actions.float().cpu().detach().numpy()) ** 2).mean().item())
         # arr[i*batch_size:(i+1)*batch_size] = actions.float().cpu().detach().numpy().astype(np.float16)
 
 arr.flush()
